zipspool: New filament guide and clip
The existing clip forced the filament to unravel towards the center of
the clip.  This caused additional friction and could cause the
filament to become crimped on its way into the filament guide.
Replace with a new clip and carriage system.  The carriage is designed
to alow the filament to pull straight from the spool.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>